+++ src/main/java/gs/mclo/api/util/package-info.java
@@ -1,0 +1,6 @@
+@ApiStatus.Internal
+@NotNullByDefault
+package gs.mclo.api.util;
+
+import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNullByDefault;